+++ checkout-ui-custom/src/templates/TVorRICAMsg.js
@@ -1,5 +1,6 @@
 const TVorRICAMsg = (message) => `
   <div id="tfg-custom-tvrica-msg" class="tfg-custom-msg">
+    <p class="tfg-custom-icon"></p>
     <p class="tfg-custom-text">${message}</p>
   </div>
 `;